+++ R/snapshot_to_reconstruction_context.R
@@ -61,7 +61,7 @@
 #' - `observation_id`;
 #' - `structural_group`;
 #' - `component`;
-#' - `record_set_id`;
+#' - `record_set_identifier`;
 #' - `resource_id`;
 #' - `locator_path`.
 #'
@@ -93,8 +93,7 @@
 #' [snapshot_to_recordset_df()],
 #' [subset_snapshot()],
 #' [add_snapshot_context()],
-#' [add_structural_groups()],
-#' [create_record_set()]
+#' [add_structural_groups()].
 #'
 #' @examples
 #' data("fscontextdemo_snapshot_01")
@@ -261,8 +260,8 @@ snapshot_to_reconstruction_context <- function(
 
 
   contextual_snapshot |>
-    create_record_set(
-      record_set_id = "structural_group",
+    record_set_projection(
+      record_set_identifier = "structural_group",
       resource_id = "inst_id",
       locator_path = "rel_root_path",
       construction_rule = c(